requirements.txt
asgiref==3.11.0
Django==5.1.15
django-tinymce==5.0.0
pillow==12.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
whitenoise==6.11.0
